+++ src/main/kotlin/model/Cuboid.kt
@@ -1,6 +1,6 @@
 package model
 
-open class Cuboid(var l: Float = 0.0f, var b: Float = 0.0f, var h: Float = 0.0f) : Figure() {
+open class Cuboid(var l: Double = 0.0, var b: Double = 0.0, var h: Double = 0.0) : Figure() {
     override fun planeArea() = 2 * (l * b)
     override fun totalArea() = 2 * (l * b + b * h + h * l)
     override fun volume() = l * b * h
